requirements.txt
django-admin-interface
pillow
asgiref
Django
django-jalali
jdatetime
sqlparse
pytz
djangorestframework_simplejwt
djangorestframework
django-filter
celery[redis]
psycopg2-binary
markdown
gunicorn
python-dotenv
django-storages
boto3
drf-spectacular
drf_spectacular_sidecar
requests